Move mapper to marquez.common pkg (#173)
* Move mapper to marquez.core.common pkg

* A mapper is a crosscutting concern, move to marquez.common pkg instead
